refactor: move module with key symbols into modules folder
